requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.4
django-allauth==65.3.0
django-bootstrap4==24.4
django-channels==0.7.0
django-cors-headers==4.6.0
django-guardian==2.4.0
django-heroku==0.3.1
django-ratelimit==4.1.0
django-storages==1.14.4
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stripe==11.3.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
